Move ApplicationLifecycle to core module
