+++ core/src/main/java/fr/nihilus/music/core/CommonModule.kt
@@ -18,12 +18,15 @@ package fr.nihilus.music.core
 
 import android.content.Context
 import android.content.SharedPreferences
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.ProcessLifecycleOwner
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import fr.nihilus.music.core.lifecycle.ApplicationLifecycle
 import fr.nihilus.music.core.os.*
 import javax.inject.Singleton
 
@@ -50,5 +53,8 @@ abstract class CommonModule {
                 "${appContext.packageName}_preferences",
                 Context.MODE_PRIVATE
             )
+
+        @Provides @ApplicationLifecycle
+        fun providesApplicationLifecycle(): LifecycleOwner = ProcessLifecycleOwner.get()
     }
 }
